Require Go 1.20 for Go Programgen
We added program generation that assumes availability of generics as part of https://github.com/pulumi/pulumi/pull/13149, but didn't bump the version of Go required in the generated go.mod.  Currently, Pulumi is supported on the supporte versions of Go, which includes Go 1.20 and 1.21, so here we bump the Go version to the minimum currently supported version, which also supports all of the code we generate today.