Add support for default env values
The providers team would like to be able to run codegen without being effected by ambient
plugins. To do this, we need to:

1. Be able to disable ambient plugins ✓
2. Be able to have provider local plugins (this PR)
3. Disable downloading external plugins from the internet.

I'm hiding this feature behind an environmental variable `$PULUMI_PLUGIN_CACHE`. To
simplify internal logic (and make the default clear to the user), I'm adding an explicit
default of `$PULUMI_HOME/plugins`.

This commit just adds support for default values.