feat(L): Support marking helpers
Adds an `L.Helper()` method that marks the caller
as a helper function.
Functions that call this will be skipped when gathering
position information about where a log message comes from.

This is especially desirable as all of testify's assertion functions
call this method if it exists.
Without this, all testify messages will start at assertions.go
versus the actual test file.