test: No more sleeps
Instead of fixed sleep durations,
set up a simple contract with test processes.

Test processes print anything to stdout when they're ready,
and exit with a zero status code if they receive a signal.
They exit with a non-zero status code if they don't receive the signal
in a reasonable amount of time.