Rewrite filestate.RenameStack logic in terms of checkpoints
The RenameStack function in filestate was one of the few places outside
of "cmd" still using `stack.DefaultSecretsProvider`, which it was only
using because the stack rename function was written in terms of
deserialised snapshots instead of serialised deployments.

This change rewrites the RenameStack function to work on the raw
deployment data rather than deserialised snapshot. This allows us to
remove the use of `stack.DefaultSecretsProvider` from filestate.

This would also allow the service to update their use of RenameStack to
work on the raw deployment as well, saving them the use of their
NoopSecretsProvider.