Tests and fix for --target-dependents with explicit providers
Fixes https://github.com/pulumi/pulumi/issues/13591.

This changes the logic for providers to always be targeted, this means
they can be skipped from --targets lists most of the time.

Because they don't need to be in the --targets list it makes the
behaviour of --target-dependents much more useful. If you want to update
a resource and it's children but it has an explicit provider you can
just --targets the resource.

If you want to use --target-dependents to target _all_ the resources
managed by an explicit provider that will work if the provider is in
--targets.