Support bailing from RunFunc
**Background**

The result.Result type is used by our CLI implementation
to communicate how we want to exit the program.

Most `result.Result` values (built from errors with `result.FromError`)
cause the program to print the message to stderr
and exit the program with exit code -1.
The exception is `result.Bail()`,
which indicates that we've already printed the error message,
and we simply need to `exit(-1)` now.

Our CLI command implementation use `cmdutil.RunResultFunc`
which takes a `func(...) result.Result` to implement this logic.

`cmdutil` additionally includes a `cmdutil.RunFunc`
which takes a `func(...) error` and wraps it in `RunResultFunc`,
relying on `result.FromError` for the conversion:

    func RunFunc(run func(...) error) func(...) {
        return RunResultFunc(func(...) result.Result {
            if err := run(...); err != nil {
                return result.FromError(err)
            }
            return nil
        })
    }

**Problem**

In CLI contexts where we're using an `error`,
and we want to print an error message to the user and exit,
it's desirable to use diag.Sink to print the message to the user
with the appropriate level (error, warning, etc.)
and exit without printing anything else.

However, the only way to do that currently is
by converting that function to return `result.Result`,
turn all error returns to `result.FromError`,
and then return `result.Bail()`.

**Solution**

This change introduces a `result.BailError` error
that gets converted into a `result.Bail()`
when it passes through `result.FromError`.

It allows commands implementations that use `error`
to continue returning errors and still provide an ideal CLI experience.

It relies on `errors.Is` for matching, so even if an intermediate layer
wraps the error with `fmt.Errorf("..: %w", ErrBail)`,
we'll recognize the request to bail.

BailError keep track of the internal error that triggered it, which
(when everything is moved off of result and onto error) means we'll
still be able to see the internal errors that triggered a bail during
debugging.

Currently debugging engine tests is pretty horrible because you often
just get back a `result.Result{err:nil}` with no information where in
the engine stack that came from.

**Testing**

Besides unit tests, this includes an end-to-end test
for using RunResultFunc with a bail error.
The test operates by putting the mock behavior in a fake test,
and re-running the test binary to execute *just that test*.

**Demonstration**

This change also ports the following commands to use BailError:
cancel, convert, env, policy rm, stack rm.

These command implementations are simple and were able to switch easily,
without bubbling into a change to a bunch of other code.