Use pulumi.Context's WorkGroup if available
If the `context.Context` given to `Compose` came from a pulummi.Context,
use that pulumi.Context's WorkGroup.
This will affect the lifetime of the `Run` or `RunErr` call
that owns the `pulumi.Context`.

To accomplish this, we put the WorkGroup on the context
with new internal functions: WithWorkGroup and GetOrCreateWorkGroup.