make(sdk/py): Fix install target
The `make install_plugin` and `make_build` targets
are now broken because pulumi-language-python is
an independent Go module,
and it has a different import path than it did before.

This updates the Makefile to match the new path
and uses `-C` to cd into that directory to run `go install`.