[sdk/go] Fix issues with templates (#13872)
Addresses two issues with codegen templates in the Go SDK:

1. `type_builtins.go.template` was including an import for a package
that doesn't exist. (The import doesn't exist in the current generated
code that is checked-in). This change removes the import line from the
template.

2. Some changes had been made directly to `config/require.go`, but this
is a generated file -- the changes should have been made to the
template. This change moves the changes to the template.

After the changes to these templates, running `make gen` in `sdk/go`
yields no diffs.

Fixes #13871