Kp/adds root pulumi ai command (#13808)
<!--- 
Thanks so much for your contribution! If this is your first time
contributing, please ensure that you have read the
[CONTRIBUTING](https://github.com/pulumi/pulumi/blob/master/CONTRIBUTING.md)
documentation.
-->

# Description

<!--- Please include a summary of the change and which issue is fixed.
Please also include relevant motivation and context. -->

This PR is intended to add the beginnings of a set of `pulumi ai`
commands to the Pulumi CLI. The first implementation is relatively
straightforward - we want to be able to open the Pulumi AI app from the
CLI while providing a simple set of operations for users - pre-filling
the AI prompt, and optionally not running that prompt when the page is
opened. The CLI should open the appropriate URL in the user's default
browser, and the app should take over from there.

This was implemented as a subcommand for `pulumi ai` as we intend to add
more features both as siblings and at the parent `pulumi ai` command in
the future, as this current functionality does not represent the "final
default" experience we'd like users to have when running `pulumi ai`.

Fixes https://github.com/pulumi/pulumi.ai/issues/126

## Checklist

- [x] I have run `make tidy` to update any new dependencies
- [x] I have run `make lint` to verify my code passes the lint check
  - [ ] I have formatted my code using `gofumpt`

<!--- Please provide details if the checkbox below is to be left
unchecked. -->
- [ ] I have added tests that prove my fix is effective or that my
feature works
<!--- 
User-facing changes require a CHANGELOG entry.
-->
- [x] I have run `make changelog` and committed the
`changelog/pending/<file>` documenting my change
<!--
If the change(s) in this PR is a modification of an existing call to the
Pulumi Cloud,
then the service should honor older versions of the CLI where this
change would not exist.
You must then bump the API version in
/pkg/backend/httpstate/client/api.go, as well as add
it to the service.
-->
- [ ] Yes, there are changes in this PR that warrants bumping the Pulumi
Cloud API version
<!-- @Pulumi employees: If yes, you must submit corresponding changes in
the service repo. -->