Prepare for v3.82.0 release (#13908)
We're planning to release a new minor version of the CLI on Monday, both
for some new features (e.g. CLI support for org search and AI) and also
so that the docker containers for v3.82.0 are released on Monday [with
Node 18](https://github.com/pulumi/pulumi-docker-containers/pull/150)
since Node 16 will be EOL. Going ahead and prepping the version bump now
with this PR. I'll open and merge the associated freeze PR on Monday
once the remaining org search PRs have been merged.