Update ExecuteRegisterResourceOutputs to use error
Part of the yak shave to get rid of the `Result` type. We need to do these
changes from the bottom up because we allow converting from `error` to
`Result` (we simply discard data), but not the other way (because we'd
have to make up an error and `from Result conversion` isn't very
informative).

This replaces the `Result` return value in
`ExecuteRegisterResourceOutputs` to `error`. This means we now get
errcheck lints which warned that we we're discarding the result from
this call in the import code.