Update ExecuteRegisterResourceOutputs to use error (#13915)
<!--- 
Thanks so much for your contribution! If this is your first time
contributing, please ensure that you have read the
[CONTRIBUTING](https://github.com/pulumi/pulumi/blob/master/CONTRIBUTING.md)
documentation.
-->

# Description

<!--- Please include a summary of the change and which issue is fixed.
Please also include relevant motivation and context. -->

Part of the yak shave to get rid of the `Result` type. We need to do
these changes from the bottom up because we allow converting from
`error` to `Result` (we simply discard data), but not the other way
(because we'd have to make up an error and `from Result conversion`
isn't very informative).

This replaces the `Result` return value in
`ExecuteRegisterResourceOutputs` to `error`. This means we now get
errcheck lints which warned that we were discarding the result from this
call in the import code.